Create tables, CRUD for provincial, sector, cluster, chapter, members

+++ migrations/m150912_154118_create_user_table.php
@@ -11,7 +11,7 @@ class m150912_154118_create_user_table extends Migration
             'id' => $this->primaryKey(),
             'auth_key' => $this->string(32)->notNull(),
             'division' => $this->string(15)->notNull(),
-            'service_id' => $this->integer(),
+            'service_id' => $this->smallInteger(),
             'email' => $this->string(255)->notNull(),
             'status' => $this->smallInteger()->notNull(),
             'role' => $this->string(15)->notNull(),
